+++ views/_footer.php
@@ -1,3 +1,21 @@
 
-</body>
+			</div><!-- .box -->
+		</section><!-- .main -->
+	</div>
+</div><!-- .cover -->
+			
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
+    <script src="assets/js/bootstrap.min.js"></script>
+    <script type="text/javascript">
+	    $(document).ready(function () {
+ 
+		window.setTimeout(function() {
+		    $(".alert").fadeTo(1500, 0).slideUp(500, function(){
+		        $(this).remove(); 
+		    });
+		}, 5000);
+		 
+		});
+	</script>
+  </body>
 </html>