+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGE LOG
 
+**June 1, 2015**
+- Added Bootstrap 3 Style
+- Added autoclosing alerts
+
 **May 31 2015**
 - Forked devplanete/php-login-advanced